[iOS SDK][Hackbook] Prevent crash in Hackbook caused by incorrect topViewController assumption
Summary:
If Hackbook receives the "application:openURL:options:" callback when MainViewController is not the topViewController it will crash since the `updateDeepLinkLabel:` method in HFAppDelegate will try to call the `updateDeepLinkLabel:` method on whatever the top view controller is which in this case was `NativeShareDialogTestViewController`.

Here is a screenshot of the issue in Xcode
{F775234227}

This crash was reported by QA by clicking on **Share** > **Share Dialog**. It would then try to re-log into Facebook (for some reason that is still unknown). The login would succeed but would crash when navigating back from the Facebook app to Hackbook. See the following for a demo:
https://www.internalfb.com/intern/px/p/2fgSl

I was initially able to replicate this on a device, however after I logged out of Hackbook and logged in again this behavior was no longer occurring, i.e. it would navigate to the "Share Dialog" page successfully without requesting to log in again. I'm still not sure why login was being triggered at that point.

I did find another way to replicate the callback to Hackbook when MainViewController was not on top to reproduce the crash itself. See the following video:
https://pxl.cl/2fqfS

I added a basic `respondsToSelector:` check to fix the unrecognized selector crash so even if login is triggered again Hackbook, should not crash.

Test Plan: Ran tests in Xcode

Reviewers: joesusnick, josuebrizuelaxp, samodom

Reviewed By: joesusnick

Differential Revision: https://phabricator.intern.facebook.com/D39909354

Tasks: T132951101

Tags: accept2ship, publish_when_ready