[iOS SDK] Convert FBSDKTokenCache.h/m to Swift
Summary:
There is an unexplained behavior difference between retrieving a cached Access vs Auth token.

For Access, failure to unarchive a token always clears the cache.
For Auth, failure to unarchive a token will not clear the cache.
In both cases it will return nil.

Additionally, this diff fixes a likely bug in retrieving an AuthenticationToken when `shouldUseTokenOptimizations` is true. If it cannot find a token in the keychain, it will clear any token in UserDefaults. However, it was using the wrong key to try and do this so it was a noop.

Test Plan: Review carefully since this type is not well unit tested.

Reviewers: samodom, #sdk

Reviewed By: samodom

Differential Revision: https://phabricator.intern.facebook.com/D39889997

Tasks: T133072546

Tags: opensource-needs-export, swift-adoption, accept2ship, publish_when_ready