[iOS SDK] Replace usage of NSMutableArray with NSArray in FBAEMReporter's internal API
Summary: A NSMutableArray was being used in cases in which it wasn't needed. Because of this in the Swift tests an explicit NSMutableArray had to be created. After updating the API to use NSArray instead of NSMutable array, a Swift array can be used directly.

Test Plan: Ran tests in Xcode

Reviewers: joesusnick, josuebrizuelaxp, samodom

Reviewed By: samodom

Differential Revision: https://phabricator.intern.facebook.com/D40089515

Tags: opensource-needs-export, swift-adoption, accept2ship, publish_when_ready