Fix private method definition inconsistency and add a few missing imports
Summary:
All of the other public "configureWithNetworker" methods have an override for NS_SWIFT_NAME that renames the Swift method to use "networker" instead of "withNetworker". This "fixes" the definition of what the private method should be in Swift.
Also adding a few missing imports that will have to be eventually added when the FBAEMKit.h file is removed.

Reviewed By: samodom

Differential Revision: D40191595

fbshipit-source-id: 033c316c42ec976702bb40aba7e074703200df68