Added Swift and temporary ObjC unit tests for FBSDKAppLinkNavigation
Summary: Added more unit tests to FBSDKAppLinkNavigation to test how the method `navigate` behavior works in both languages. The plan is to remove the Objc methods once we have converted AppLinkNavigation to Swift, and we ensure it continues to work as before in both languages or to discover how different the behavior is after the conversion.

Reviewed By: samodom

Differential Revision: D40164976

fbshipit-source-id: 3bd73e535fb13833cfd4083f55032a5579629d2c