[iOS SDK] Added Swift and temporary ObjC unit tests for FBSDKAppLinkNavigation
Summary: Added more unit tests to FBSDKAppLinkNavigation to test how the method `navigate` behavior works in both languages. The plan is to remove the Objc methods once we have converted AppLinkNavigation to Swift, and we ensure it continues to work as before in both languages or to discover how different the behavior is after the conversion.

Test Plan: sandcastle

Reviewers: samodom, #sdk

Reviewed By: samodom

Differential Revision: https://phabricator.intern.facebook.com/D40164976

Tags: opensource-needs-export, swift-adoption, accept2ship, publish_when_ready