[iOS SDK] Fix OTA installs for iOS 16
Summary:
This is a bit confusing.

Trying to install on iOS 16 results in the error:
```
-[IXSErrorPresenter presentErrorForIdentities:code:underlyingError:errorSource:]: Asked to present alert for error 1 source MobileInstallation underlying error Error Domain=MIInstallerErrorDomain Code=180 "The item being installed did not contain an app with bundle ID "com.facebook.unknown", but it did contain app(s) with bundle ID(s): {(
    "com.facebook.internal.sdk.samples.Hackbook.internalInHouse"
)}" UserInfo={SourceFileLine=450, NSLocalizedDescription=The item being installed did not contain an app with bundle ID "com.facebook.unknown", but it did contain app(s) with bundle ID(s): {(
    "com.facebook.internal.sdk.samples.Hackbook.internalInHouse"
)}, FunctionName=-[MIInstaller _bundleInDirectory:withBundleID:error:]} for identities (
    "[com.facebook.unknown/P]"
)
```

It seems this has to do with OTA installs and how they use a manifest.plist file to specify the bundle identifier. This is generated in https://fburl.com/code/15pbvgu2 and uses the `package` key of the manifest info that’s stored on the IEntMobileOrSmokeBuild - pointer: https://fburl.com/code/vot4hppl.

In theory setting the key in this file, and then using it in D40321851

Test Plan: Land this, land a www change, see if signing is fixed.

Reviewers: michaelyan45, ldelgado, samodom, #sdk

Reviewed By: samodom

Differential Revision: https://phabricator.intern.facebook.com/D40321894

Tags: accept2ship, publish_when_ready