[iOS SDK] Bump waitForExpectations from 2 to 4 in attempt to fix test flakyness
Summary: Might as well try to wait a bit longer if it prevents the test failing.

Test Plan:
Ran tests in Xcode

I also did a quick test and it looks like the NSPredicate check happens every 1 second so if there is something that delayed this for an extra second, that might be causing the flakyness.

{F785178611}

Reviewers: joesusnick, josuebrizuelaxp, samodom

Reviewed By: samodom

Differential Revision: https://phabricator.intern.facebook.com/D40648425

Tasks: T132421535

Tags: opensource-needs-export, accept2ship, publish_when_ready