[CAPIG]Optimize the config loading
Summary:
We optimize the CAPIG config loading in this diff. We don't load the CAPIG config every time we log CAPIG event, but only load the config if:
1. Last refresh timestamp is over 24 hours, or
2. CAPIG config doesn't exist (e.g. loaded config is not valid)

We also add a flag to mark whether we're loading config now, if yes, then we just add the callback to the callback list and wait for the config loading completion. In order to make those thread-safe, we create a serial queue and put the operations inside the serial queue.

Test Plan:
1. Unit Test
2. Launch CoffeeShop and CAPIG events logging work as expected

Reviewers: jajiang, #ads_sdk

Reviewed By: jajiang

Subscribers: xinwu330

Differential Revision: https://phabricator.intern.facebook.com/D45035238

Tags: opensource-needs-export