[coffee shop] fix a crash when clicking the UITree button
Summary:
The `FBSDKCodelessIndexer` class does not have this method `recursiveCaptureTree:`. Simply check it before calling it.

Not sure why we have that button there, a proper fix would be to figure out what we want to capture there (can be a follow-up diff), but let's prevent the app from crashing first.

Test Plan:
Before:

 https://pxl.cl/30rgr

After:

 https://pxl.cl/30rq1

Reviewers: xinwu330, #mobile_sdk

Reviewed By: xinwu330

Differential Revision: https://phabricator.intern.facebook.com/D47570565