Replaces unnecessary usage of mach_absolute_time
Summary:
Under the new Apple guidelines for required reason apis, `mach_absolute_time` will need to be listed in the privacy manifest. The Apple documentation suggests using `clock_gettime_nsec_np` as an alternative: https://developer.apple.com/documentation/kernel/1462446-mach_absolute_time.

https://developer.apple.com/documentation/bundleresources/privacy_manifest_files/describing_use_of_required_reason_api

"""
System boot time APIs
The following APIs for accessing the system boot time require reasons for use. Use the string NSPrivacyAccessedAPICategorySystemBootTime as the value for the NSPrivacyAccessedAPIType key in your NSPrivacyAccessedAPITypes dictionary.

systemUptime

**mach_absolute_time()**

In your NSPrivacyAccessedAPITypeReasons array, supply the relevant values from the list below.

35F9.1
Declare this reason to access the system boot time in order to measure the amount of time that has elapsed between events that occurred within the app or to perform calculations to enable timers.

Information accessed for this reason, or any derived information, may not be sent off-device. There is an exception for information about the amount of time that has elapsed between events that occurred within the app, which may be sent off-device.
"""

Reviewed By: ppansy

Differential Revision: D48069727

fbshipit-source-id: 8ffc8146d5516870765d1acf8e86af85737db6ae