[iOS][HackBook][ez] Improving Logging Style in Console
Summary:
# Context

I am trying to use SKAN debugger in HackBook app, but the console styling is a bit awkward. So I wanted to change it, and get familiarized with the iOS SDK diff workflow.

# What I changed

Just changed the way we printed to console. It is now more readable with newline symbol.

Test Plan:
Before:

 {F1463476232}

After:

 {F1463475775}

Reviewers: ishandabholkar, jajiang, xinwu330, #app_signals, #app_aem_and_platform_api

Reviewed By: ishandabholkar

Differential Revision: https://phabricator.intern.facebook.com/D54436730

Tags: ez_review