[FBSDK][PrivacyManifest] SPM Dynamic Fix
Summary: The binary of built dynamic frameworks contains some folders which are just symbolic links to other folders. When we zip these using ```zip -r```, the symbolic links are converted to hard links. This is not only an invalid build but also, because this ```PackageManifestGeneration``` command occurs after signing, this creates an invalid signature. In order to preserve the symbolic links, we update this command to ```zip -r --symlinks```.

Test Plan:
Ran Sandcastle job:
https://www.internalfb.com/sandcastle/job/36028798240815720

Verified the signatures locally.

Reviewers: jajiang, taox, #ads_sdk, #privacy_manifest

Reviewed By: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D55052410