[T186510880] rename projects.git.vip to git.vip
Summary:
Details in the attached task.
This is for git http migration.
This url doesn't work with the migration mechanism, but the other one should work just as fine as it's just a CNAME to the same thing as the other one.

Test Plan: This url doesn't work with the migration mechanism, but the other one should work just as fine as it's just a CNAME to the same thing as the other one.

Reviewers: singhsrb

Reviewed By: singhsrb

Subscribers: #pesourcecontrol

Differential Revision: https://phabricator.intern.facebook.com/D57437506

Tasks: T186510880

Tags: opensource-needs-export