[FBSDK] Access Token Enhancement
Summary:
Because FBLogin sets ```currentAccessToken``` to an invalid access token for opt-out, logged in users, we want to explicitly set the token string of the tokenless, ads related endpoints to nil.

We will also not use ```currentAccessToken``` for batch requests when the token is likely invalid.

Test Plan:
- Ran Unit Tests
- Verified all ads related endpoints are tokenless via the graph API explorer
- Verified no access token errors for ads related endpoints in both Hackbook and Coffee Shop
- Verified changes did not break non-ads endpoints

Reviewers: jajiang, #ads_sdk

Reviewed By: jajiang

Subscribers: #adnw

Differential Revision: https://phabricator.intern.facebook.com/D57139188

Tags: opensource-needs-export