Move away from Asset Library to Photos local asset identifiers by default
Summary: Asset Library framework has been deprecated for a while and will be removed "in a future major release" of iOS. For this reason we are migrating to local asset identifier for videos provided by Photos framework

Test Plan:
This doesn't work with the current version of FB. This diff is required: D60691152. So with this diff applied we can test that sending local asset identifier works.
{F1788179704}

Reviewers: adamcmiel, ryantobin, yuchilee, #swift_wilde, #mobile_sdk, #creation_growth_ios

Reviewed By: yuchilee

Subscribers: leejun

Differential Revision: https://phabricator.intern.facebook.com/D60787998

Tags: opensource-needs-export