Lay foundation for IAP Unit Tests
Summary:
One critical component of implementing the IAP enhancement is writing effective unit tests. IAP unit tests operate a little differently than regular FBSDK unit tests. They depend on the ```SKTestSession``` class provided by the ```StoreKitTest``` library. The APIs in ```SKTestSession``` require the tests to have a "host app" in order to work correctly. Because this is an SDK, we do not have one by default. Thus, this diff lays the foundation for IAP unit tests by doing the following:
1. Create a new target called ```IAPTestsHostApp``` that will serve as the host app for the IAP tests
2. Create a new testing target called ```IAPTests``` whose host app is the ```IAPTestsHostApp``` target. This target will hold all of the IAP unit tests that depend on ```SKTestSession```
3. Add these targets to the ```BuildAllKits``` and ```FBSDKCoreKit``` schemes
4. Create a ```StoreKitTestCase``` class that configures the ```SKTestSession```. All IAP tests can inherit from this.
5. Create the ```FBSDKStoreKitConfigurationUnitTests.storekit``` configuration file, which will hold the test products, subscriptions, etc. we can use in our test cases.

Reviewed By: jjiang10

Differential Revision: D61809245

fbshipit-source-id: fce9325e347d1e2ad9c97349b8137862e9724d34