Create the transaction observer
Summary: One of the critical components of the IAP implementation is the transaction observer, which is used to observe both restored transactions and new candidate transactions. After observing these transactions, the transaction observer calls the transaction logger, which will ultimately decide if we should log the transaction and log it if so. This diff implements the ```IAPTransactionObserver```.

Reviewed By: jjiang10

Differential Revision: D62038383

fbshipit-source-id: fbcbc10fcc44d5a20d5e66c4a68a99870d82da80