[FBSDK iOS][IAP][4/N] Create the transaction observer
Summary: One of the critical components of the IAP implementation is the transaction observer, which is used to observe both restored transactions and new candidate transactions. After observing these transactions, the transaction observer calls the transaction logger, which will ultimately decide if we should log the transaction and log it if so. This diff implements the ```IAPTransactionObserver```.

Test Plan:
- Created unit tests in next diff
- Ran unit tests
- CI / CD

Reviewers: jajiang, yifan0726, shawnzeng, #ads_sdk, maxalbright

Reviewed By: jajiang

Subscribers: taox

Differential Revision: https://phabricator.intern.facebook.com/D62038383

Tags: opensource-needs-export