Create the event resolver
Summary:
One of the critical components of the IAP implementation is the event resolver. After the ```IAPTransactionObserver``` observes a new transaction, it calls the transaction logger. Before the transaction logger can log a high quality event based on the transaction, it must gather all the necessary data (i.e. event name, product title, product id, etc.) about that transaction.

The transaction logger will call the event resolver to resolve all of this information. It will receive this information back in the form of an ```IAPEvent``` object. After this information is resolved, the transaction logger can log an event. This diff implements the ```IAPEventResolver```.

Reviewed By: jjiang10

Differential Revision: D62085759

fbshipit-source-id: 7112b0283c53bbd51cb650b5a9c3eddf63cdbbd0