Add additional subscription info
Summary:
The Legacy Store Kit implementation logs some additional subscription information in the parameters of the event such as the subscription period, whether or not the subscription has an introductory offer, the subscription period of that introductory offer, the price of that introductory offer, and whether or not the subscription offers a free trial.

To stay consistent, we will also log this information in our Store Kit 2 implementation. This diff updates our logic to also gather this information.

Reviewed By: jjiang10

Differential Revision: D62197861

fbshipit-source-id: af6370994adac9842d66d887a76cc09981ac6287