[FBSDK iOS] Add UI tests for UIApplication.open API
Summary:
The UIApplication.open API was broken in XCode 16 and Swift only and it worked
as expected in Objective-C. There is more context in D63081488.
This added UI tests for UIApplication.open API specifically in swift to cover
the APIs for both browser and deeplink use cases.

Test Plan:
- Verify web URL can be opened succesfully without warnings in browser when built against XCode 16.
- Verify app deeplink URL can be opened successfully without warnings when built against XCode 16
  and app is installed.

Reviewers: xinwu330, kylinchang, #mobile_sdk

Reviewed By: xinwu330

Differential Revision: https://phabricator.intern.facebook.com/D63204401

Tasks: T202205217

Tags: xcode, uikit-diff