[FBSDK iOS][IAP][16/N] Update transaction id to be a string
Summary: We are integrating Store Kit 1 with Store Kit 2. In Store Kit 1, the transaction ID is a string while in Store Kit 2, it is a UInt64. While the types are different, that values will be the same. In our legacy implementation, we passed the transaction id as a string in the event payload. We need to be consistent. This diff updates our use of transaction id to consistently be a string.

Test Plan:
- Ran unit tests
- CI / CD

Reviewers: jajiang, taox, #ads_sdk

Reviewed By: jajiang, taox

Subscribers: taox

Differential Revision: https://phabricator.intern.facebook.com/D62599386