Update transaction id to be a string
Summary: We are integrating Store Kit 1 with Store Kit 2. In Store Kit 1, the transaction ID is a string while in Store Kit 2, it is a UInt64. While the types are different, that values will be the same. In our legacy implementation, we passed the transaction id as a string in the event payload. We need to be consistent. This diff updates our use of transaction id to consistently be a string.

Reviewed By: jjiang10, xta0

Differential Revision: D62599386

fbshipit-source-id: 5caeabb75a45ce96e058d7335839c6aba3e4268d