Update IAPEventResolver to support Store Kit 1
Summary:
In order to integrate Store Kit 1 with our Store Kit 2 implementation, we need the ```IAPEventResolver``` to be able to resolve events for transactions that come from Store Kit 1 or Store Kit 2. This diff updates the ```IAPEventResolver``` implementation to support Store Kit 1.

After observing a transaction, we will then resolve the event using the appropriate ```IAPEventResolver``` API depending on which version of Store Kit the transaction was observed from.

Reviewed By: jjiang10

Differential Revision: D62814661

fbshipit-source-id: 6f9e5afe2defe7a65e17fb3b6f93592a871d18d1