[FBSDK iOS][IAP][18/N] Update IAPEventResolver to support Store Kit 1
Summary:
In order to integrate Store Kit 1 with our Store Kit 2 implementation, we need the ```IAPEventResolver``` to be able to resolve events for transactions that come from Store Kit 1 or Store Kit 2. This diff updates the ```IAPEventResolver``` implementation to support Store Kit 1.

After observing a transaction, we will then resolve the event using the appropriate ```IAPEventResolver``` API depending on which version of Store Kit the transaction was observed from.

Test Plan:
- Added unit tests in next few diffs
- Ran unit tests
- CI / CD

Reviewers: jajiang, shawnzeng, #ads_sdk

Reviewed By: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D62814661