[FBSDK iOS][IAP][31/N] Update candidate date to latest transaction purchase date
Summary:
We use a sliding window technique to limit the transaction history we consider when observing / logging transactions. In the ```IAPTransactionCache```, we have a variable called ```newCandidatesDate```; upon each transaction observation iteration, we only consider transactions that are newer than this ```newCandidatesDate```.

Previously, after observing the new candidate transactions, we would update ```newCandidatesDate``` to ```now```. Theoretically, after we observe the new candidate transactions but before ```newCandidatesDate``` is updated to ```now```, a new transaction could occur. With this logic, we would miss this new transaction in the next observation.

This diff fixes this edge case by updating ```newCandidatesDate``` to the purchase date of the last observed candidate transaction as opposed to ```now```.

Test Plan:
- Updated unit tests
- Ran unit tests
- CI / CD

Reviewers: liuxulei, jajiang, xinwu330, bozhu, #ads_sdk

Reviewed By: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D62969430