[FBSDK iOS][IAP][32/N] Make IAPEvent isSubscription a stored property
Summary:
Previously, if the subscription GK was turned off, we did not log subscriptions at all; however, in the next diff, we update this logic to log a purchase event for subscription purchases when the subscription GK is turned off (this is consistent with the [legacy store kit 1 implementation](https://www.internalfb.com/code/facebook-ios-sdk/[main%3A549428142bd1454724f8f5a45988525f45222c72]/FBSDKCoreKit/FBSDKCoreKit/AppEvents/Internal/FBSDKPaymentProductRequestor.m?lines=124-133)).

As a prerequisite, we need to make the ```isSubscription``` property on the ```IAPEvent``` struct a stored property whose value is based on the transaction or product rather than a computed property whose value is based on the event name. This is because a subscription could be resolved as a purchase event when the subscription GK is turned off, and we still need to log the subscription parameters for this purchase event.

Test Plan:
- Updated unit tests
- Ran unit tests
- CI / CD

Reviewers: jajiang, #ads_sdk

Reviewed By: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D62975462