Set new candidate date to now on first observation
Summary:
After the SDK is updated with the new IAP implementation, the first time the app launches, the ```newCandidatesDate``` from the ```IAPTransactionCache``` will be nil, which means we will observe and log the entire transaction history of the user up to this point. However, we should only log new transactions that occur after the update or we risk duplicate logging transactions that were already logged from the legacy implementation.

This diff sets the ```newCandidatesDate``` value to ```now``` on the first app launch so that we only consider new transactions that occur after the update and avoid the potential duplicate logging.

Reviewed By: jjiang10

Differential Revision: D62988700

fbshipit-source-id: 0ca5c0186d59b82d237237d47d36a2f67be71f80