[FBSDK iOS][IAP][38/N] Configure IAP Observation Time From Server
Summary:
The IAP observation time refers to how frequently we check a user's transaction history for new transactions in nanoseconds. The default is 3600000000000 nanoseconds or 1 hour. In D63005744, we updated the app events config API to include the ```ios_iap_observation_time```.

This diff updates the app events configuration on the client-side to read and store this value. We then set the IAP observation time based on the cached app events configuration.

Test Plan:
### Automated Tests
- Updated unit tests
- Ran unit tests
- CI / CD

### Manual Tests
- Set SDK Version to 17.2.0
- Ran Coffee Shop
- Set a break point and ensured we read the observation time from the server:
 {F1877664082}
- Set a break point and ensured we set it correctly in the ```IAPTransactionObserver```:
 {F1877666832}

Reviewers: jajiang, #ads_sdk

Reviewed By: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D63052228