Configure IAP Observation Time From Server
Summary:
The IAP observation time refers to how frequently we check a user's transaction history for new transactions in nanoseconds. The default is 3600000000000 nanoseconds or 1 hour. In D63005744, we updated the app events config API to include the ```ios_iap_observation_time```.

This diff updates the app events configuration on the client-side to read and store this value. We then set the IAP observation time based on the cached app events configuration.

Reviewed By: jjiang10

Differential Revision: D63052228

fbshipit-source-id: 0535f88e148da197cf1178653c19c753d7f20c54