[FBSDK][IAP][7/N] Fix decimal point issue
Summary:
We observed that sometimes when logging the value of a transaction, it was logged with an excess number of decimal points (i.e. 4.99 was 4.9900000000000002 or 1.99 was 1.9899999999999998). This can hurt the event quality.

The root cause is because doubles do not have as much precision as decimals and when converting the price from a decimal to a double, we lost precision. To fix it, we first round the number to 3 decimal places (this is the most any currency has) and then we log it as an ```NSDecimalNumber``` as opposed to a ```Double``` to maintain precision.

Test Plan:
- CI / CD
- Ran Coffee Shop and autologged an IAP event. Observed the value to sum in the payload had the correct number of decimal places:
```
{
  "event": {
    "fb_transaction_id": "4",
    "fb_description": "Purchase a Coffee Shop Rewards Card",
    "fb_iap_product_type": "inapp",
    "fb_content_title": "Rewards Card",
    "fb_transaction_date": "2024-10-10 02:32:26+0000",
    "fb_num_items": 1,
    "_logTime": 1728527548,
    "_implicitlyLogged": "1",
    "_valueToSum": 4.99,
    "fb_iap_sdk_supported_library_versions": "SK1-SK2",
    "is_autolog_app_events_enabled": "1",
    "_eventName": "fb_mobile_purchase",
    "fb_currency": "USD",
    "is_implicit_purchase_logging_enabled": "1",
    "fb_content_id": "com.facebook.coffeeshop.nonconsumable.rewardscard",
    "fb_iap_client_library_version": "SK1"
  },
  "isImplicit": true
}
```

Reviewers: jajiang, #ads_sdk

Reviewed By: jajiang

Subscribers: taox

Differential Revision: https://phabricator.intern.facebook.com/D64154628

Tags: opensource-needs-export

Privacy Context Container: L1285343