Test manual purchase failed API
Summary: Unlike in store kit 1, we cannot observe and autolog purchase failed events in store kit 2. Thus, we provide an API that developers can call if they would like to preserve the logging of purchase failed events. The previous diff D64256380 created this API; this diff adds tests for it.

Reviewed By: jjiang10

Differential Revision: D64266042

fbshipit-source-id: 72a07dd7a1370a4fc393b9dea477cdeb8a90d426