Fetch and Parse Dedupe Window from Server
Summary:
To implement autolog and manual log dedupe, we need to configure the "dedupe window" from the server to give us more flexibility in case we need to extend or reduce the window. We will only dedupe events that occur within the given dedupe window. We pass this window down as a part of the app events config endpoint.

This diff fetches and parses the "dedupe window" from app events config endpoint and stores the result in the ```FBSDKAppEventsConfiguration```.

Reviewed By: jjiang10

Differential Revision: D64566047

fbshipit-source-id: 240a45a31a43a670a20941cd2c46eedc50c4c279