[FBSDK][IAP][16/N] Fetch and Parse Dedupe Window from Server
Summary:
To implement autolog and manual log dedupe, we need to configure the "dedupe window" from the server to give us more flexibility in case we need to extend or reduce the window. We will only dedupe events that occur within the given dedupe window. We pass this window down as a part of the app events config endpoint.

This diff fetches and parses the "dedupe window" from app events config endpoint and stores the result in the ```FBSDKAppEventsConfiguration```.

Test Plan:
### Unit Tests
- Added unit tests
- CI / CD

### Manual Tests
- Ran coffee shop and added a breakpoint where we parse the dedupe window
- Ensured the dedupe window passed from server was the same as what we parsed and stored on the client-side:
 {F1936362616}
 {F1936362786}

Reviewers: jajiang, kongxinzhu, #ads_sdk, maxalbright

Reviewed By: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D64566047

Tags: opensource-needs-export

Privacy Context Container: L1285343