Save Potential Duplicate Events When App Enters Background
Summary:
In D64674786, we implemented dedup for manually logged and implicitly logged IAP events. It is possible that the app user terminates or backgrounds the app before we have fully processed the potential duplicate events; if this happens, we risk losing the events.

To avoid this, when the app enters the background, we cache the potential duplicate events. Then, when the app is reopened, we process them.

Reviewed By: jjiang10

Differential Revision: D64712984

fbshipit-source-id: 63fc3e95b372e0a55641ed90a474a7d356e18cae