[FBSDK][IAP][20/N] Save Potential Duplicate Events When App Enters Background
Summary:
In D64674786, we implemented dedup for manually logged and implicitly logged IAP events. It is possible that the app user terminates or backgrounds the app before we have fully processed the potential duplicate events; if this happens, we risk losing the events.

To avoid this, when the app enters the background, we cache the potential duplicate events. Then, when the app is reopened, we process them.

Test Plan:
- Added unit tests and did manually testing in the next diff
- CI / CD

Reviewers: jajiang, taox, #ads_sdk

Reviewed By: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D64712984

Tags: opensource-needs-export

Privacy Context Container: L1285343