More Dedup Testing
Summary:
In D64712984, we implemented caching unprocessed potential duplicate events when the user backgrounds the app and processing them when the user brings the app back to the foreground.

This diff adds unit tests for this functionality. This diff also adds manually logging for IAP events in Coffee Shop so we can manually test the dedup implementation.

Reviewed By: jjiang10

Differential Revision: D64716200

fbshipit-source-id: a9842fa3d08c39a4107dd004fcb5127290e41338