[FBSDK][IAP][23/N] Only verify verifiable events
Summary: To be consistent with the legacy implementation, we should only send a validation result for events that are validated (i.e. purchases, subscriptions, and restored events). We also update the default quantity to be 1 when the transaction is nil because we cannot have a purchase for a quantity of 0.

Test Plan: - CI / CD

Reviewers: jajiang, #ads_sdk

Reviewed By: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D65005873

Tags: opensource-needs-export

Privacy Context Container: L1285343