Only verify verifiable events
Summary: To be consistent with the legacy implementation, we should only send a validation result for events that are validated (i.e. purchases, subscriptions, and restored events). We also update the default quantity to be 1 when the transaction is nil because we cannot have a purchase for a quantity of 0.

Reviewed By: jjiang10

Differential Revision: D65005873

fbshipit-source-id: bd31adf9622357e68135ae1fd6fe09c7f7e6d477