[FBSDK iOS] Migrate Hackbook from deprecated openURL method
Test Plan:
<Summary might be visible to public on our GitHub repo page but test plan is stripped>

## Summary

Hackbook app is broken on **iOS 18** when testing Reels and Stories sharing. The reason for that is removal of implementation of deprecated openURL API:
https://fb.workplace.com/groups/777581527394095/posts/984800950005484

To unblock developers debugging and QA team testing FB iOS SDK via Hackbook we need to migrate it to modern variant which is available since iOS 10. It's safe to do so because both Hackbook and FB iOS SDK have iOS 12 as a minimum version

## Test plan

|Destination|Before|After|
|--
|Story| https://pxl.cl/5R76N |  https://pxl.cl/5R772
|Reels| https://pxl.cl/5R76W |  https://pxl.cl/5R779

Reviewers: ryantobin, zifeip, #mobile_sdk, #creation_growth_ios

Reviewed By: zifeip

Subscribers: leejun

Differential Revision: https://phabricator.intern.facebook.com/D65232659

Tags: bugfix, uikit-diff