Dedup Optimization
Summary: In the original dedup implementation, we only checked for implicit transactions at the beginning of the dedup window when a manual transaction was logged. However, based on testing, an implicit transaction could occur or be "finished" some elapsed time after the manual transaction occurs but still within the dedup window. Therefore, we update the implementation to recheck for implicit transactions again at the end of the dedup window so we don't miss this edge case.

Reviewed By: jjiang10

Differential Revision: D65046779

fbshipit-source-id: c2f180c3007929abb5a6cebe5884d7284ca4085f