[FBSDK][IAP] Dedup Optimization
Summary: In the original dedup implementation, we only checked for implicit transactions at the beginning of the dedup window when a manual transaction was logged. However, based on testing, an implicit transaction could occur or be "finished" some elapsed time after the manual transaction occurs but still within the dedup window. Therefore, we update the implementation to recheck for implicit transactions again at the end of the dedup window so we don't miss this edge case.

Test Plan:
- CI / CD
- Updated coffee shop to "finish" a transaction after manually logging it; observed we were still able to dedup it

Reviewers: jajiang, taox, #ads_sdk, maxalbright

Reviewed By: jajiang

Subscribers: taox

Differential Revision: https://phabricator.intern.facebook.com/D65046779

Tags: opensource-needs-export