Drop Duplicate Initiated Checkout Events
Summary: In store kit 2, we currently send initiated checkout events anytime a purchase happens or a purchase failure happens. Because some purchases that are sent are actually restored purchases (i.e. we transform them on the server-side), we should not send an initiated checkout event for these purchases. On the client-side, however, we do not yet know these purchases are actually restored purchases so we log an initiated checkout event. This diff updates the client-side logic to send the necessary information so that we can check if this is a duplicate initiated checkout event on the server-side and drop it if so.

Reviewed By: jjiang10

Differential Revision:
D66025394

Privacy Context Container: L1285343

fbshipit-source-id: 6e8c21a79da1f259e68c26e0c399283658fd45c0