[FBSDK][IAP][1/N] Drop Duplicate Initiated Checkout Events
Summary: In store kit 2, we currently send initiated checkout events anytime a purchase happens or a purchase failure happens. Because some purchases that are sent are actually restored purchases (i.e. we transform them on the server-side), we should not send an initiated checkout event for these purchases. On the client-side, however, we do not yet know these purchases are actually restored purchases so we log an initiated checkout event. This diff updates the client-side logic to send the necessary information so that we can check if this is a duplicate initiated checkout event on the server-side and drop it if so.

Test Plan:
### Unit Tests
- CI / CD

### Manual Tests
- Ran Coffee Shop connected to my OD
- Sent 2 duplicated initiated checkout events
- Slogged the events before and after IAP processing:

**Event 1 Before**:
{F1962394971}

**Event 1 After:**
{F1962395211}

**Event 2 Before**:
{F1962395748}

**Event 2 After:**
This event was dropped as expected

Reviewers: jajiang, kongxinzhu, #ads_sdk, maxalbright

Reviewed By: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D66025394

Privacy Context Container: L1285343