[FBSDK][IAP+Integrity][3/N] Support operational parameters in App Events State
Summary: This diff adds support for the operational parameters in ```FBSDKAppEventsState```.

Test Plan:
### Internal Summary
For full context, please see the [FBSDK IAP Integrity Design & Technical Plan](https://docs.google.com/document/d/1IlHeQ5M65KhEB4bY-fk-0hS-zsJoU3H2ZhS5KJS6F8c/) doc. At a high level, there are often times that the SDK needs to send some event parameters that are used for operational, debugging, or processing purposes. The SDK currently has no way to differentiate these special parameters from ```custom_data``` parameters that are intended to be set by the advertiser. This leads to 1 of 2 problems: either a) these parameters are filtered out by integrity or b) we allowlist these parameters but then they end up downstream in offsite signals when they should not.

To solve this problem, we define a new field for ```operational_parameters``` on the AE-API and App CAPI schema that enables us to send these operational parameters separate from custom data. D65565467 builds this functionality on the server-side.

In the app events flow on the client-side, we hold the events that will be flushed in the ```FBSDKAppEventsState``` object. This diff supports adding the operational parameters to this object.

### Test Plan
- CI / CD

Reviewers: liuxulei, jajiang, #ads_sdk, maxalbright

Reviewed By: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D66025391

Tags: opensource-needs-export

Privacy Context Container: L1285343