[FBSDK][IAP][9/N] Expose Transaction Cache to Objective-C
Summary:
During testing, we noticed an edge case where if the app has autologging turned off and then re-enables, all of a user's transaction history will be logged, which would likely result in over reporting of events. We should assume that these transactions have already been handled and only consider the transactions that occur after the advertiser enables autologging. To achieve this, we set the new candidates date to ```now``` whenever we observe that autologging is turned off. In doing this, we will only consider transactions that have occurred after this date.

The first step in implementing this is exposing the transaction cache to objective-c, which is done in this diff.

Test Plan: - CI / CD

Reviewers: jajiang, kylinchang, #ads_sdk

Reviewed By: jajiang

Subscribers: maxalbright

Differential Revision: https://phabricator.intern.facebook.com/D66303119

Privacy Context Container: L1285343