Set new candidates date when autologging is off
Summary:
During testing, we noticed an edge case where if the app has autologging turned off and then re-enables it, all of a user's transaction history will be logged, which would likely result in over reporting of events. We should assume that these transactions have already been handled and only consider the transactions that occur after the advertiser enables autologging. To achieve this, we set the new candidates date to ```now``` whenever we observe that autologging is turned off. In doing this, we will only consider transactions that have occurred after this date.

This diff implements this functionality.

Reviewed By: jjiang10

Differential Revision: D66309371

fbshipit-source-id: b578b5782e8b4ead3bc58569a3bc3d9c17b1762e