[FBSDK][IAP][10/N] Set new candidates date when autologging is off
Summary:
During testing, we noticed an edge case where if the app has autologging turned off and then re-enables it, all of a user's transaction history will be logged, which would likely result in over reporting of events. We should assume that these transactions have already been handled and only consider the transactions that occur after the advertiser enables autologging. To achieve this, we set the new candidates date to ```now``` whenever we observe that autologging is turned off. In doing this, we will only consider transactions that have occurred after this date.

This diff implements this functionality.

Test Plan:
### Unit Tests
- Updated unit tests
- CI / CD

### Manual Tests
- Ran coffee shop with autologging turned off
- Made a purchase in Store Kit 2
- Observed that we never sent a purchase event for it, as expected
- Turned autologging back on
- Made a purchase in Store Kit 2
- Observed that we sent a purchase event for it, as expected

Reviewers: jajiang, #ads_sdk, maxalbright

Reviewed By: jajiang

Subscribers: jajiang

Differential Revision: https://phabricator.intern.facebook.com/D66309371

Tags: opensource-needs-export

Privacy Context Container: L1285343