Dedup Edge Case Fix
Summary:
There is a dedup edge case where if a purchase was manually logged and the app was backgrounded before the corresponding IAP event was observed and before the dedup processor processed it, we would miss deduping the duplicate events.

This diff fixes this case by ensuring the dedup processor always checks a user's transaction history after reopening the app and processing saved events that have not been processed yet.

Reviewed By: jjiang10

Differential Revision: D66941530

fbshipit-source-id: 8e666e78523981005d85dd2cafe341f224d07d41