Purchase Restored Bug Fix
Summary:
Apple's documentation specifically mentions that non-renewable subscriptions and consumables are not restorable transactions. As a result, our previous Store Kit 1 implementation does not support restoring these types of purchases. Our new implementation should match this behavior. This is fixed in 2 ways:
1. We do not allow logging these purchase restored events directly from the client-side
2. We append the "product type" as an operational parameter in the event payload so that if we transform this event to be restored on the server-side, we drop it.

This diff implements this for the SDK side.

Reviewed By: jjiang10

Differential Revision: D66970013

fbshipit-source-id: 780b69e15509e44a7889054cbece792448e64204