Add Product ID to Transaction Cache
Summary:
There is an edge case where a subscription can have different "tiers" or "levels" of subscription. If a user upgrades their subscription "level", it will share the same "original transaction id" as the original lower tier subscription. Because we only log a subscription event once as opposed to every time it renews (and use the "original transaction id" to ensure this), this prevents us from logging a new subscription event when the user upgrades their subscription "tier"; we want to capture this event.

To fix this, this diff adds the product id as an attribute of the transaction cache, ensuring that we don't block new subscription products that share the same transaction id as other subscription products.

Reviewed By: jjiang10

Differential Revision: D66985280

fbshipit-source-id: e3d0b1ef31591201e2175cd4cba4e7d7b31db55c