Fix price for consumables
Summary:
Per [Apple's documentation](https://developer.apple.com/documentation/storekit/transaction/price), the ```transaction.price``` attribute should account for the ```purchasedQuantity``` of consumables and encompass the total amount spent on the IAP. However, after testing, we noticed this isn't the case and it always reflects the price of a single unit. This may be a bug on Apple's side; however, this diff updates the logic to ensure we always reflect the correct price.

If a transaction is a consumable, has a purchased quantity greater than 1, and ```transaction.price``` equals ```product.price``` (meaning the transaction price reflects a single unit), we update the amount to equal ```product.price * transaction.purchasedQuantity```.

Differential Revision: D68128501

fbshipit-source-id: dc3fff780f682d37ab4004fcd358434bbeff346a