Fix Legacy SKAN Bug for Fine Conversions
Summary:
For each SKAN conversion rule, there are a series of events. After **all** of the events have occurred for a given conversion rule, the value of that conversion rule will be sent to Apple by means of the SKAN APIs.

While investigating the implementation of SKAN in FBSDK, we came across this bug which, in some cases, will return that a rule has been matched if the first event in the rule matches a recorded event even if the following events would not match. Instead of early returning, we need to make sure that all events in a given conversion rule match the event history before hitting the SKAN APIs.

Reviewed By: KylinChang

Differential Revision: D74680353

fbshipit-source-id: 6622f9ab7a48129334cce8cb20ec6f3cc2ff15fb