Fix Legacy SKAN Bug for Coarse Conversions (SKAN 4)
Summary:
For each SKAN conversion rule, there are a series of events. After **all** of the events have occurred for a given conversion rule, the value of that conversion rule will be sent to Apple by means of the SKAN APIs.

While investigating the implementation of SKAN in FBSDK, we came across this bug which, in some cases, will return that a rule has been matched if the first event in the rule matches a recorded event even if the following events would not match. Instead of early returning, we need to make sure that all events in a given conversion rule match the event history before hitting the SKAN APIs.

The previous diff updated the logic for fine conversions; we need to do the same for coarse conversions in SKAN 4.

Reviewed By: KylinChang

Differential Revision: D74682014

fbshipit-source-id: 5633fbb3e18ba34d44f5b019e6a526f0cf7c0e1a