banned params manager will send banned param names -iOS SDK
Summary:
# context
https://docs.google.com/document/d/1v0lzxH1qK5L-mmzmSFf-eb3j2z8sACKV7m1AWAdUnns/edit?tab=t.fci6ojfeiv6z#heading=h.sfrjokyrejjs
(See ParamChange.1 - Standard Parameter Filtering)
We'd like to start transmitting the names of standard parameters that we block as part of our tiered enforcement treatment. This is consistent with what we already do for client side custom parameter filtering.
We have legal alignment + LAMA to do this
# this diff

Implement the straightforward change in the ios SDK. See D83934924 for the analogous android diff.

Reviewed By: maxalbrightmeta

Differential Revision: D84034108

fbshipit-source-id: d71a5a251dcaba8fe94b2591cb7ba67f5be54ee1