Stop Clearing External ID Each Time the Advertiser Sets User Data
Summary:
# External Summary

The external id is a field that is set by advertisers to uniquely identify a user.

This is a bug wherein the API we surface to advertisers to set all fields at once does calls the UserDataStore with a ```nil``` externalID. The way the logic is written in the data store means that this will actually clear the persisted external id.

Instead of doing that, we should set the external id value to be equivalent to the existing value. If the advertiser wants it to be cleared, they can call the clearUserData API.

Reviewed By: ryantobinmeta, Zhiyi-Zhang

Differential Revision: D86901688

fbshipit-source-id: 6c7352b7da6a45a48b4a34b976ad69ca1945cd1c