[FB SDK][Bug Fix] Stop Clearing External ID Each Time the Advertiser Sets User Data
Summary:
# External Summary

The external id is a field that is set by advertisers to uniquely identify a user.

This is a bug wherein the API we surface to advertisers to set all fields at once does calls the UserDataStore with a ```nil``` externalID. The way the logic is written in the data store means that this will actually clear the persisted external id.

Instead of doing that, we should set the external id value to be equivalent to the existing value. If the advertiser wants it to be cleared, they can call the clearUserData API.

Test Plan:
# Internal Summary
# Context

The external id is a field that is set by advertisers to uniquely identify a user. On the backend, once there is a strong identity match for an event with a given external id, we can use the field therafter to perform the same match, even in the absence of other stronger match keys.

This is a bug that also existed in the Ads SDK wherein the API we surface to advertisers to set all fields at once does calls the UserDataStore with a ```nil``` externalID. The way the logic is written in the data store means that this will actually clear the persisted external id.

Instead of doing that, we should set the external id value to be equivalent to the existing value. If the advertiser wants it to be cleared, they can call the clearUserData API or the
```
- (void)setUserData:(nullable NSString *)data
            forType:(EnigmaUserDataType)type
```

API with a nil value.

# This Diff

In D86738603, we fixed this for the ads sdk. In this diff, we introduce the fix for the FBSDK.

{F1983473951}

Rollback Plan:

Reviewers: ryantobin, zhiyi, kylinchang

Reviewed By: ryantobin, zhiyi

Differential Revision: https://phabricator.intern.facebook.com/D86901688

Tags: opensource-needs-export